Separate the two asset generators and support choosing one at build time
